# Add `Memory::alloc_static_zeroed` to allocate memory that's filled with zeroes.