This is generally faster than `malloc` followed by `memset`. It's essentially the `Memory` equivalent of `calloc`.

The reason is that pages handed out by the OS are usually zeroed out already. This is for security purposes, as you cannot risk handing out traces of memory used by another application. The OS will do this in its spare time.
Most allocators make use of this in `calloc`. If large amounts of memory are requested, it will just request zeroed pages from OS, and can hand them back immediately. If small amounts are requested, the allocator will just use `memset` to clear out the requested memory.

This can be exploited for performance benefit for e.g. `resize_zeroed` and `is_zero_constructible` structs: If we request zeroed pages for these calls, we can omit our own `memset` calls, which will make the allocation near instantaneous.