@aaronp64 2025-03-14
https://github.com/godotengine/godot/pull/91633 looks similar, adding `Memory::calloc_static`.  Just taking a quick look through, I like the idea of using `p_ensure_zero` to reuse most of the function if we can, not sure if there's other differences between the two.

For naming I'm leaning more toward `calloc_static`, but either sounds good to me.