@Ivorforce 2025-03-14
> https://github.com/godotengine/godot/pull/91633 looks similar

This is the same idea indeed!

I think calling the function `calloc_static` is somewhat dangerous: `calloc` normally accepts 2 parameters. If you call `Memory::calloc_static(1, 512)`, analogously to `calloc(1, 512)`, you'd get 1 byte with `p_pad_align=true`. This is because `Memory::calloc_static` accepts parameters `bytes` and `pad_align`, while `calloc` accepts `num` and `bytes`.

I instead opted to call it `zeroed` because it's consistent with other functions with similar behavior, and because (I think) it's more obvious what it does to people who have never heard of `calloc`.