@DeeJayLSP 2025-04-16
Wouldn't it be fair to add a `memcalloc()` or `memalloc_zeroed()` here?
https://github.com/godotengine/godot/blob/3c0652699add1ca69ef18dc4408e044f485ab829/core/os/memory.h#L111-L115