@Ivorforce 2025-04-26
Thanks a lot @DeeJayLSP!
I've patched up all the spots you mentioned. Might as well make this a performance PR to motivate it better!