@Ivorforce 2025-04-26
I benchmarked a ~4% improvement for `HashMap` insertion:
```c++
	// Setup
	HashMap<int, int> hash_map;

	auto t0 = std::chrono::high_resolution_clock::now();
	for (int i = 0; i < 20000000; i ++) {
		// Test
		hash_map.insert(i, i);
	}
	auto t1 = std::chrono::high_resolution_clock::now();
	std::cout << std::chrono::duration_cast<std::chrono::milliseconds>(t1 - t0).count() << "ms\n";
```

This printed:
- `2789ms` on `master`
- `2681ms` on this PR