@Ivorforce 2025-05-12
I rebased, and ran a slightly more complicated test (to cover more cases):

```c++
	for (int size : { 1, 2, 8, 64, 1024, 4096}) {
		auto t0 = std::chrono::high_resolution_clock::now();

		for (int run = 0; run < 20000000 / size; run++) {
			HashMap<int64_t, int64_t> dictionary;
			for (int idx = 0; idx < size; idx ++) {
				// Test
				dictionary.insert(idx, idx);
			}
		}
		auto t1 = std::chrono::high_resolution_clock::now();

		std::cout << "size:" << size << std::endl;
		std::cout << std::chrono::duration_cast<std::chrono::milliseconds>(t1 - t0).count() << "ms\n";
	}
```

This printed on master:
```
size:1
1464ms
size:2
944ms
size:8
579ms
size:64
709ms
size:1024
980ms
size:4096
1328ms
```

On this pr:
```
size:1
1268ms
size:2
838ms
size:8
550ms
size:64
672ms
size:1024
841ms
size:4096
1149ms
```

This new test suggests that, at least for fundamental types in key/value, `HashMap` insertion can be 15% faster (for normal sizes anyway).