@Ivorforce 2025-05-18
Out of interest, I just tested again using `memset` instead of `alloc_static_zeroed` on HashMap. I got this:

master:
```
size:1
1430ms
size:2
921ms
size:8
563ms
size:64
670ms
size:1024
817ms
size:4096
1134ms
```

pr:
```
size:1
1283ms
size:2
838ms
size:8
545ms
size:64
652ms
size:1024
806ms
size:4096
1088ms
```

It's the same test, but quite different results from last time for `master`, so this definitely is a bit flaky.
In any case, it does seem like `memset` covers _most_ of the performance benefit, but not _all_ of it. I would therefore recommend to move forwards with this PR rather than detouring over `memset` first.